+++ app/packages/core/src/components/Filters/categoricalFilter/filterOption/FilterItem.tsx
