+++ app/packages/core/src/components/Filters/utils.ts
@@ -1,5 +1,9 @@
+import * as fos from "@fiftyone/state";
+import { genSort } from "../../utils/generic";
+
 const NONSTRING_VALUES: any[] = [false, true, null];
 const STRING_VALUES = ["False", "True", "None"];
+export const CHECKBOX_LIMIT = 20;
 
 export const getValueString = (value): [string, boolean] => {
   if (NONSTRING_VALUES.includes(value)) {
@@ -36,4 +40,29 @@ export const joinStringArray = (arr: string[]) => {
     result += `and ${arr[arr.length - 1]}`;
     return result;
   }
+};
+
+export const nullSort = ({
+  count,
+  asc,
+}: fos.SortResults): ((aa: V, bb: V) => number) => {
+  return ({ count: aac, value: aav }, { count: bbc, value: bbv }): number => {
+    let a = [aav, aac];
+    let b = [bbv, bbc];
+
+    if (count) {
+      a.reverse();
+      b.reverse();
+    }
+
+    let result = 0;
+    for (let i = 0; i < a.length; i++) {
+      result = genSort(a[i], b[i], asc);
+      if (result !== 0) {
+        return result;
+      }
+    }
+
+    return result;
+  };
 };
